+++ packages/client/src/App.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { onMount } from "svelte";
-  import { S, joinByCode, openHelp, initMotion } from "./lib/store.svelte";
+  import { S, joinByCode, openHelp, initMotion, initDebug } from "./lib/store.svelte";
   import Home from "./lib/Home.svelte";
   import Lobby from "./lib/Lobby.svelte";
   import Game from "./lib/Game.svelte";
@@ -8,9 +8,11 @@
   import Connecting from "./lib/Connecting.svelte";
   import Help from "./lib/Help.svelte";
   import EventOverlay from "./lib/EventOverlay.svelte";
+  import Debug from "./lib/Debug.svelte";
 
   onMount(() => {
     initMotion();
+    initDebug();
     const r = new URLSearchParams(location.search).get("r");
     if (r) {
       joinByCode(r);
@@ -42,3 +44,4 @@
 {#if S.reconnecting}<div class="reconnect-overlay"><div class="rc-box">Reconnecting...</div></div>{/if}
 {#if S.connecting}<Connecting />{/if}
 {#if S.showHelp}<Help />{/if}
+{#if S.debug}<Debug />{/if}